requirements.txt
amqp==5.3.1
anyio==4.8.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.5
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-celery-results==2.5.1
django-redis==5.4.0
drf-yasg==1.21.8
flower==2.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
loguru==0.7.3
openai==1.59.8
packaging==24.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg==3.2.4
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
